final_version ready for vogsphere

+++ srcs/requirements/wordpress/tools/create_wordpress.sh
@@ -1,20 +1,25 @@
 #!/bin/sh
 
-#Download wordpress
-wget https://wordpress.org/latest.tar.gz
-tar -xzvf latest.tar.gz
-rm -rf latest.tar.gz
+if [ -f ./wordpress/wp-config.php ]
+then
+	echo "wordpress already downloaded"
+else
+	#Download wordpress
+	wget https://wordpress.org/latest.tar.gz
+	tar -xzvf latest.tar.gz
+	rm -rf latest.tar.gz
 
-#Update configuration file
-rm -rf /etc/php/7.3/fpm/pool.d/www.conf
-mv ./www.conf /etc/php/7.3/fpm/pool.d/
+	#Update configuration file
+	rm -rf /etc/php/7.3/fpm/pool.d/www.conf
+	mv ./www.conf /etc/php/7.3/fpm/pool.d/
 
-#Inport env variables in the config file
-cd /var/www/html/wordpress
-sed -i "s/username_here/$MYSQL_USER/g" wp-config-sample.php
-sed -i "s/password_here/$MYSQL_PASSWORD/g" wp-config-sample.php
-sed -i "s/localhost/$MYSQL_HOSTNAME/g" wp-config-sample.php
-sed -i "s/database_name_here/$MYSQL_DATABASE/g" wp-config-sample.php
-mv wp-config-sample.php wp-config.php
+	#Inport env variables in the config file
+	cd /var/www/html/wordpress
+	sed -i "s/username_here/$MYSQL_USER/g" wp-config-sample.php
+	sed -i "s/password_here/$MYSQL_PASSWORD/g" wp-config-sample.php
+	sed -i "s/localhost/$MYSQL_HOSTNAME/g" wp-config-sample.php
+	sed -i "s/database_name_here/$MYSQL_DATABASE/g" wp-config-sample.php
+	mv wp-config-sample.php wp-config.php
+fi
 
 exec "$@"
